Keep event stores and read repositories in separate packages
