+++ readrepository/memory/readrepository.go
